elf: Move Elf64FileRange to separate module
In the effort of splitting up the ELF parser code, move the
Elf64FileRange implementation to its own module.
